Create Operator Overloading
The following program follows a tutorial from the following site:
http://www.raywenderlich.com/80818/operator-overloading-in-swift-tutorial
Further features and functionality of operator overloading and creating can be found there.
Johnny pointed out that a prime example of using an operator overload is when using a two dimensional array, adding two points together (like (1, 2) and (5, 3)) to get a third point (6, 5). Useful in math and games using this layout.